requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.5
Django==5.0.4
django-rest-passwordreset==1.4.0
djangorestframework==3.15.1
idna==3.6
kombu==5.3.6
ndg-httpsclient==0.5.1
pip-system-certs==4.0
prompt-toolkit==3.0.43
pyasn1==0.6.0
pycparser==2.22
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
rabbitmq==0.2.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
wrapt==1.16.0
